% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_axes.R
\name{add_axes}
\alias{add_axes}
\title{add axes and ticks to an svg document.}
\usage{
add_axes(g_id, axes, fig)
}
\description{
add axes and ticks to an svg document.
}
\details{
transformation from plot space to pixel space is handled here.
}

