% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_pt.R
\name{dim_coords}
\alias{dim_coords}
\title{svg coordinates that are side agnostic}
\usage{
dim_coords(vals, val.lims, svg.lims, log = FALSE, as.crd = TRUE)
}
\arguments{
\item{vals}{coordinates}

\item{val.lims}{coordinate limits}

\item{svg.lims}{svg view pane limits}

\item{log}{boolean for logged side (only \code{FALSE} is supported currently)}

\item{as.crd}{boolean for passing result to \code{as.crd}}
}
\description{
translates values into svg space
}
\seealso{
\code{\link{svg_coords}}
}
\keyword{internal}

