% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_svg.R
\name{svg}
\alias{svg}
\title{render svg}
\usage{
svg(object, ...)
}
\description{
render svg
}
\examples{
\dontrun{
library(gsplot)
gs <- gsplot()
gsNew <- points(gs, y=1:10, x=2:11, 
            col="blue", pch=18, hovertext='point')
svg(gsNew)
}
}

