% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_pt.R
\name{svg_coords}
\alias{svg_coords}
\title{svg coordinates from x,y coordinates}
\usage{
svg_coords(x, y = NULL, xlim, ylim = NULL, view.bounds, log = "",
  as.crd = TRUE)
}
\arguments{
\item{x}{location of x coordinate}

\item{y}{location of y coordinate}

\item{xlim}{limits of x coordinate reference}

\item{ylim}{limits of y coordinate reference}

\item{view.bounds}{a named vector of \code{x},\code{y},\code{width}, and \code{height}}

\item{log}{a character of '', 'x', or 'xy' (only '' is currently supported)}

\item{as.crd}{boolean for passing result to \code{as.crd}}
}
\description{
translates coordinate reference space into svg space.
}
\seealso{
\code{\link{dim_coords}}
}

