% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_svg.R
\name{svg_points}
\alias{svg_points}
\title{points to svg}
\usage{
svg_points(svg, window, x, y, col, cex = 1, ...)
}
\arguments{
\item{\dots}{all end up at attributes to the element}
}
\description{
points to svg
}

