% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_abline.R
\name{render_abline}
\alias{render_abline}
\title{create abline svg elements}
\usage{
render_abline(g.view, a = NULL, b = NULL, h = NULL, v = NULL,
  reg = NULL, coef = NULL, untf = FALSE, lty = par("lty"),
  col = par("col"), lwd = par("lwd"), xlim, ylim, ...)
}
\description{
creates abline elements using the syntax of \code{\link[graphics]{abline}}
}
\examples{
\dontrun{
library(gsplot)
gs <- gsplot() \%>\% 
   points(y=1:11, x=1:11, 
            col="blue", pch=18, xlab='pizza', ylab='dogs') \%>\% 
   abline(h=1:4, v=5:8)
svg(gs)
}
}

