% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_svg.R
\name{svg}
\alias{svg}
\title{render svg}
\usage{
svg(object, ...)
}
\description{
render svg
}
\examples{
\dontrun{
library(gsplot)
gs <- gsplot() \%>\% 
   points(y=1:10, x=2:11, 
            col="blue", pch=18, hovertext=paste0('text:',1:10)) \%>\% 
   points(x=4:11, y=11:4, 
            col="red", pch=1, hovertext=paste0('text:',11:4))
svg(gs)
}
}

