% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_svg.R
\name{svg}
\alias{svg}
\title{render svg}
\usage{
svg(object, ...)
}
\description{
render svg
}
\examples{
\dontrun{
library(gsplot)
gs <- gsplot() \%>\% 
   points(y=1:11, x=1:11, 
            col="blue", pch=18, hovertext=paste0('text:',1:11), xlab='pizza', ylab='dogs') \%>\% 
   points(x=4:11, y=11:4, 
            col="red", pch=1, hovertext=paste0('text:',11:4)) \%>\% 
   points(3:5,4:6,side=c(1,4), col='green', hovertext='green', pch=23, ylab='cats') \%>\% 
   lines(2:5, c(2,2.6,2.3), col='blue')
svg(gs)
}
}

